+++ deprecated_files/resources/scripts/choose_new_pixels_to_scan.py
@@ -1,1 +1,0 @@
-../../python/choose_new_pixels_to_scan.py