+++ resources/launch_scripts/docker/launch_client.sh
@@ -32,7 +32,7 @@ def extract_opt_path(py_args, opt):
 
 py_args, debug_dir = extract_opt_path(py_args, "--debug-directory")
 py_args, gcd = extract_opt_path(py_args, "--gcd-dir")
-py_args, startup = extract_opt_path(py_args, "--startup-json-dir")
+py_args, startup = extract_opt_path(py_args, "--client-startup-json")
 
 dockermount_args = ""
 py_args += " "
@@ -46,8 +46,8 @@ if gcd:
     # NOTE: WE ARE NOT FORWARDING THIS ARG TO THE SCRIPT B/C ITS PASSED WITHIN THE STARTUP.JSON
     #
 if startup:
-    dockermount_args += f"--mount type=bind,source={startup},target=/local/startup-json-dir "
-    py_args += f"--startup-json-dir /local/startup-json-dir "
+    dockermount_args += f"--mount type=bind,source={os.path.dirname(startup)},target=/local/startup "
+    py_args += f"--client-startup-json /local/startup/{os.path.basename(startup)} "
 
 print(f"{dockermount_args}#{py_args}")
 ')
@@ -74,6 +74,9 @@ docker run --network="host" $pull_policy --rm -i \
     --env PY_COLORS=1 \
     $(env | grep '^SKYSCAN_' | awk '$0="--env "$0') \
     $(env | grep '^EWMS_' | awk '$0="--env "$0') \
+    $(env | grep '^RABBITMQ_' | awk '$0="--env "$0') \
+    --env "RABBITMQ_HEARTBEAT=${RABBITMQ_HEARTBEAT:-600}" \
+    $(env | grep '^PULSAR_' | awk '$0="--env "$0') \
     --env "PULSAR_UNACKED_MESSAGES_TIMEOUT_SEC=${PULSAR_UNACKED_MESSAGES_TIMEOUT_SEC:-900}" \
     icecube/skymap_scanner:${SKYSCAN_DOCKER_IMAGE_TAG:-"latest"} \
     python -m skymap_scanner.client \